api/requirements.txt
Flask==3.0.3
rembg==2.0.59
pillow==10.4.0
werkzeug==3.0.4
gunicorn==23.0.0
flask-cors==5.0.0